[apps/statistics/graph] Remove a C-style cast

+++ apps/statistics/graph/box_controller.cpp
@@ -65,7 +65,7 @@ bool BoxController::reloadBannerView() {
   int precision = Preferences::sharedPreferences()->numberOfSignificantDigits();
   Poincare::Preferences::PrintFloatMode displayMode = Poincare::Preferences::sharedPreferences()->displayMode();
   // With 7 = KDFont::SmallFont->glyphSize().width(), should fit in one line
-  constexpr static int k_bufferSize = 1 + Ion::Display::Width / 7 - (int)sizeof("V1/N1");
+  constexpr static int k_bufferSize = 1 + Ion::Display::Width / 7 - sizeof("V1/N1");
   char buffer[k_bufferSize] = "";
 
   // Display series name
